Grand splitting + renaming
